Adding casecomonenet to CMQE tests and moving manual tests
