+++ cfme/tests/containers/test_basic_metrics.py
@@ -67,6 +67,7 @@ def test_basic_metrics(provider):
     Polarion:
         assignee: juwatts
         caseimportance: medium
+        casecomponent: Containers
         initialEstimate: 1/6h
     """
     try:
@@ -90,6 +91,7 @@ def test_validate_metrics_collection_db(provider,
     Polarion:
         assignee: juwatts
         caseimportance: medium
+        casecomponent: Containers
         initialEstimate: 1/6h
     """
     assert provider.wait_for_collected_metrics(
@@ -105,6 +107,7 @@ def test_validate_metrics_collection_provider_gui(appliance, provider,
     Polarion:
         assignee: juwatts
         caseimportance: medium
+        casecomponent: Containers
         initialEstimate: 1/6h
     """
     view = navigate_to(provider, "Details")
@@ -125,6 +128,7 @@ def test_flash_msg_not_contains_html_tags(provider):
     Polarion:
         assignee: juwatts
         caseimportance: medium
+        casecomponent: Containers
         initialEstimate: 1/6h
     """
     edit_view = navigate_to(provider, 'Edit')
@@ -152,6 +156,7 @@ def test_typo_in_metrics_endpoint_type(provider):
     Polarion:
         assignee: juwatts
         caseimportance: medium
+        casecomponent: Containers
         initialEstimate: 1/6h
     """
