+++ cfme/tests/containers/test_breadcrumbs.py
@@ -48,6 +48,7 @@ def test_breadcrumbs(provider, appliance, soft_assert):
     Polarion:
         assignee: juwatts
         caseimportance: medium
+        casecomponent: Containers
         initialEstimate: 1/6h
     """
     for data_set in TESTED_OBJECTS:
