+++ cfme/tests/containers/test_logging.py
@@ -74,6 +74,7 @@ def test_external_logging_activated(provider, appliance, test_item, kibana_loggi
     Polarion:
         assignee: juwatts
         caseimportance: medium
+        casecomponent: Containers
         initialEstimate: 1/6h
     """
     test_collection = ([provider] if test_item.obj is ContainersProvider
