+++ cfme/tests/containers/test_overview_data_integrity.py
@@ -67,6 +67,7 @@ def test_containers_overview_data_integrity(appliance, soft_assert):
     Polarion:
         assignee: juwatts
         caseimportance: medium
+        casecomponent: Containers
         initialEstimate: 1/6h
     """
     view = navigate_to(ContainersOverview, 'All')
