+++ cfme/tests/containers/test_provider_configuration_menu.py
@@ -30,6 +30,7 @@ def test_edit_selected_containers_provider(provider):
     Polarion:
         assignee: juwatts
         caseimportance: medium
+        casecomponent: Containers
         initialEstimate: 1/6h
     """
     '''Testing Configuration -> Edit... button functionality
@@ -54,6 +55,7 @@ def test_ocp_operator_out_of_the_box(appliance):
     Polarion:
         assignee: juwatts
         caseimportance: medium
+        casecomponent: Containers
         initialEstimate: 1/6h
     """
 
@@ -87,6 +89,7 @@ def test_pause_and_resume_provider(provider):
     Polarion:
         assignee: juwatts
         caseimportance: medium
+        casecomponent: Containers
         initialEstimate: 1/6h
     """
