+++ cfme/tests/containers/test_reload_button_provider.py
@@ -18,6 +18,7 @@ def test_reload_button_provider(provider):
     Polarion:
         assignee: juwatts
         caseimportance: medium
+        casecomponent: Containers
         initialEstimate: 1/6h
     """
