+++ cfme/tests/containers/test_start_page.py
@@ -48,6 +48,7 @@ def test_start_page(appliance, soft_assert):
     Polarion:
         assignee: juwatts
         caseimportance: medium
+        casecomponent: Containers
         initialEstimate: 1/6h
     """
     for data_set in data_sets:
