+++ cfme/tests/containers/test_static_custom_attributes.py
@@ -58,6 +58,7 @@ def test_add_static_custom_attributes(add_delete_custom_attributes, provider):
     Polarion:
         assignee: juwatts
         caseimportance: medium
+        casecomponent: Containers
         initialEstimate: 1/6h
     """
 
@@ -81,6 +82,7 @@ def test_edit_static_custom_attributes(provider):
     Polarion:
         assignee: juwatts
         caseimportance: medium
+        casecomponent: Containers
         initialEstimate: 1/6h
     """
 
@@ -109,6 +111,7 @@ def test_delete_static_custom_attributes(add_delete_custom_attributes, request, 
     Polarion:
         assignee: juwatts
         caseimportance: medium
+        casecomponent: Containers
         initialEstimate: 1/6h
     """
 
@@ -151,6 +154,7 @@ def test_add_attribute_with_empty_name(provider):
     Polarion:
         assignee: juwatts
         caseimportance: medium
+        casecomponent: Containers
         initialEstimate: 1/6h
     """
     with pytest.raises(APIException):
@@ -170,6 +174,7 @@ def test_add_date_attr_with_wrong_value(provider):
     Polarion:
         assignee: juwatts
         caseimportance: medium
+        casecomponent: Containers
         initialEstimate: 1/6h
     """
     ca = CustomAttribute('nondate', "koko", 'Date')
@@ -189,6 +194,7 @@ def test_edit_non_exist_attribute(provider):
     Polarion:
         assignee: juwatts
         caseimportance: medium
+        casecomponent: Containers
         initialEstimate: 1/6h
     """
     ca = choice(ATTRIBUTES_DATASET)
@@ -216,6 +222,7 @@ def test_delete_non_exist_attribute(provider):
     Polarion:
         assignee: juwatts
         caseimportance: medium
+        casecomponent: Containers
         initialEstimate: 1/6h
     """
     ca = choice(ATTRIBUTES_DATASET)
@@ -232,6 +239,7 @@ def test_add_already_exist_attribute(provider):
     Polarion:
         assignee: juwatts
         caseimportance: medium
+        casecomponent: Containers
         initialEstimate: 1/6h
     """
     ca = choice(ATTRIBUTES_DATASET)
@@ -251,6 +259,7 @@ def test_very_long_name_with_special_characters(request, provider):
     Polarion:
         assignee: juwatts
         caseimportance: medium
+        casecomponent: Containers
         initialEstimate: 1/6h
     """
     ca = CustomAttribute(get_random_string(1000), 'very_long_name', None)
@@ -266,6 +275,7 @@ def test_very_long_value_with_special_characters(request, provider):
     Polarion:
         assignee: juwatts
         caseimportance: medium
+        casecomponent: Containers
         initialEstimate: 1/6h
     """
     ca = CustomAttribute('very long value', get_random_string(1000), None)
