+++ cfme/tests/test_manual.py
@@ -778,31 +778,6 @@ def test_embed_tower_playbooks():
 
 
 @pytest.mark.manual
-def test_cmqe_delete_existing_resource_quota():
-    """
-    Delete existing Resource Quota
-
-    Polarion:
-        assignee: juwatts
-        caseimportance: medium
-        initialEstimate: 1h
-        title: CMQE - Delete Existing Resource Quota
-        testSteps:
-            1. &nbsp;Delete the&nbsp;existing Resource Quota that was added in&nbsp;&nbsp;
-            2. Refresh the CFME appliance
-            3. Check the containers_quota_items table in the database
-            4. Navigate to the namespace in the CFME GUI.
-        expectedResults:
-            1. Verify the Resource Quota was deleted successfully
-            2. Appliance refreshed successfully.
-            3. Verify the deleted Resource Quotas should still exist, with
-               `deleted_on` field set.
-            4. Verify the GUI displays no data referencing Resource Quota.
-    """
-    pass
-
-
-@pytest.mark.manual
 def test_osp_vmware60_test_vm_migration_with_really_long_name_upto_64_chars_worked_not_65_char():
     """
     OSP: vmware60-Test VM migration with really long name(Upto 64 chars
@@ -2630,43 +2605,6 @@ def test_osp_test_flavors_can_be_selected_creating_migration_plan():
         startsin: 5.10
         subcomponent: OSP
         title: OSP: Test flavors can be selected creating migration plan
-    """
-    pass
-
-
-@pytest.mark.manual
-def test_cmqe_modify_existing_resource_quota():
-    """
-    Modify Existing Quota
-
-    Polarion:
-        assignee: juwatts
-        caseimportance: medium
-        initialEstimate: 1h
-        title: CMQE - Modify Existing Resource Quota
-        testSteps:
-            1. Modify the existing Resource Quota that was added in &nbsp;
-            2. Refresh the CFME appliance
-            3. Check the containers_quota_items table in the database
-            4. Navigate to the namespace in the CFME GUI.
-            5. Modify the existing Resource Quota again
-            6. Refresh the CFME appliance
-            7. Check the containers_quota_items table in the database
-            8. Navigate to the namespace in the CFME GUI.
-        expectedResults:
-            1. Verify the update was successful
-            2. Appliance refreshed successfully.
-            3. Verify the modified values (eg. increased cpu) should appear
-               as multiple rows — old value with `deleted_on` field set,
-               and new value with approximately same `created_at`.
-            4. Verify the GUI displays the correct data and the fractional
-               values are formatted properly.
-            5. Verify the update was successful
-            6. Appliance refreshed successfully.
-            7. Verify multiple values with both (created_at, deleted_on)
-               set and only last one with deleted_on=null.
-            8. Verify the GUI displays the correct data and the fractional
-               values are formatted properly.
     """
     pass
 
@@ -6122,39 +6060,6 @@ def test_embed_tower_invisible():
         caseimportance: medium
         initialEstimate: 1/12h
         startsin: 5.8
-    """
-    pass
-
-
-@pytest.mark.manual
-def test_cmqe_add_new_resource_quota():
-    """
-    Add new resource quota
-
-    Polarion:
-        assignee: juwatts
-        caseimportance: medium
-        initialEstimate: 1h
-        title: CMQE - Add New Resource Quota
-        testSteps:
-            1. On the Openshift provider, create different namespaces and
-               assign them Resource Quotas. Assign different fractional
-               values to the CPU quotas like  `1.152` or `0.087` or `87m`.
-            2. Refresh the CFME appliance
-            3. Check the containers_quota_items table in the database
-            4. Navigate to the namespace in the CFME GUI.
-            5. Refresh the CFME appliance
-            6. Check the containers_quota_items table in the database
-        expectedResults:
-            1. Resource Quota assigned successfully.
-            2. Appliance refreshed successfully.
-            3. Database in `container_quota_items` table should contain
-               both old and current values.            Newly added quotas
-               should have `created_at` field set.
-            4. Verify the GUI displays the correct data and the fractional
-               values are formatted properly.
-            5. Appliance refreshed successfully.
-            6. Verify duplicate records were not added to the table
     """
     pass
